Restructured all the flask python files
