+++ SukiScan/__init__.py
@@ -1,0 +1,7 @@
+from flask import Flask
+
+app = Flask(__name__)
+app.secret_key = "SukiScan"
+
+from SukiScan import route
+from SukiScan import forms
